requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
asyncpg==0.29.0
pgvector==0.2.5
openai==1.30.1
numpy==1.26.4
python-dotenv==1.0.1
pydantic==2.7.1
